add multi-gene query support
this commit changes how parameters are handled for qmd docs. Passing more than one gene as a list cannot be accepted as a parameter. Params are key:value pairs. And a value cannot be a list. So I flattened the input$query and input$content into a character vector of length 1, I pass this to a json array as a key:value pair, that is sent to SQS, as before,  pass this list (with the flat query/content) to the report as a parameter. In the set-up chunk of the report, I take the params and turn it into a list that I call `input_report`.