adds gene report white list
this commit generates gene report whitelists, that allow anyone to get the private report for a set of pre-defined genes. A new generate_public_reports.R script defines the genes, uploads an Rds file of those genes, then generates the reports and uploads to S3. Also updates the logic/control flow of the reports to accommodate a whitelist check